Fix for cursor jumping to beginning of tinyMCE field when you focues the textarea tinymce field or when you press enter.
See the demo page (https://jsfiddle.net/matfish2/fjurzoLj/) to reproduce the unwanted behavior:
Error scenario #1. After initial page load, cick the WYSIWYG field and start typing 'abcd'. This will result in 'bcda'
Error scenario #2. In the WYSIWYG editor type a text and hit enter. The cursor will jump to te beginning of the editor field. This is unwanted behavior.

The fix will only reset the content of the tinymce field when the synced value actually changes outside of the editor from your parent Vue component.
It's a string comparison (not optimal) but it fixes the above issues.

Thanks for your library, regards, Frank